bench 测量方法学严肃修复 (do_bench 同款) + FA 对齐 fp16 + 回归 P15
方法学 (对齐 triton testing.do_bench, 对照官方实现逐条核对):
P1 ★输入轮换破 L2: 连续 forward 同一批张量, 工作集<192MB(L2) 时后 N 次全 L2 命中 → 测到 L2 带宽虚高; Ascend 无清 L2 API → n_buf 组输入轮换 (组数×工作集>L2) 等效 clear_cache
P2 多窗口 median: 原单窗口包 N 次 ÷N 只有 1 个样本 → n_rep 个独立 Event 对取 median (另报 min/mean, 抗单次抖动)
P3 时间预算自适应: 固定 warmup=5/measure=30 对快 kernel 太少 → 先 5 次估时长, warmup 25ms/rep 100ms 折算次数
P5 口径声明: 工业级=torch 全流程(含 host 调度), 我们 verify=triton 纯 kernel → json 记 time_us_min/mean + note, 文档同步

实现: bench_common.measure_event 公共测量; bench_industrial forward 改带参 fwd(*bufs[i]) (torch.compile 编译成单图) + cann-fused 同改 + 删死代码 _build_fn/_run_loop; 11 个 bench_pytorch_*.py 全部改轮换+多窗口; flash_attention 输入对齐 fp16 (工业级 FA 即 fp16, 原 fp32 vs fp16 不可比) + sweep_blocks FA setup 同步 + verify 参考升 fp32; README/ARCHITECTURE_DESIGN 测量口径同步; _sim_fix_regression.py 新增 P15 (measure_event 统计正确性 mock 单测 + 12 脚本 --help 新参数回归), 31 项全过

+++ triton_agent_optimizer/analyzers/sweep_blocks.py
@@ -483,7 +483,7 @@ def _generate_matmul_runner(op_dir: Path, candidates: List[Tuple],
         # ── Tensor setup (flash_attention) — ★布局与 kernel 一致: Q/V[nh,seq,dim], K[nh,dim,seq] ──
         seq, nh, dim = {seq}, {nh}, {dim}
         scale = 1.0 / (dim ** 0.5)
-        DTYPE = torch.float32
+        DTYPE = torch.float16   # ★与 kernel_op.py 对齐 (fp16 输入, 工业级 FA 同口径)
         device = torch.device("npu")
         q_t = (torch.randn(seq, nh, dim, dtype=DTYPE, device=device) * 0.1).permute(1, 0, 2).contiguous()
         k_t = (torch.randn(seq, nh, dim, dtype=DTYPE, device=device) * 0.1).permute(1, 2, 0).contiguous()
